ref: move schema to core (#133)
