+++ src/mcpm/commands/list.py
@@ -9,8 +9,8 @@ from rich.console import Console
 from mcpm.clients.client_config import ClientConfigManager
 from mcpm.clients.client_registry import ClientRegistry
 from mcpm.commands.server_operations.common import determine_scope
+from mcpm.core.schema import ServerConfig
 from mcpm.profile.profile_config import ProfileConfigManager
-from mcpm.schemas.server_config import ServerConfig
 from mcpm.utils.display import print_client_error, print_server_config
 from mcpm.utils.scope import ScopeType, format_scope
