+++ src/mcpm/commands/server_operations/custom.py
@@ -5,7 +5,7 @@ from rich.console import Console
 from rich.prompt import Confirm, Prompt
 
 from mcpm.commands.server_operations.common import client_add_server, determine_scope, profile_add_server
-from mcpm.schemas.server_config import SSEServerConfig, STDIOServerConfig
+from mcpm.core.schema import SSEServerConfig, STDIOServerConfig
 from mcpm.utils.display import print_server_config
 from mcpm.utils.scope import ScopeType
